+++ src/app/(stack)/store/[id]/review/page.tsx
@@ -1,88 +1,25 @@
 "use client";
 
-import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { useParams, useRouter, useSearchParams } from "next/navigation";
-import { useState } from "react";
-import { createReview, updateReview } from "@/entities/review/api/reviewApi";
-import { KEYWORD_MAP, REVIEW_FILTERS } from "@/entities/review/model/constants";
+import { useReviewForm } from "@/features/review/lib/hook/useReviewForm";
 import { ConfirmModal } from "@/features/review/ui";
 import { FilterSection, InputReview, TransitionLayout } from "@/shared/ui";
 import { CTA } from "@/shared/ui/CTA";
 
-const REVERSE_KEYWORD_MAP: { [key: string]: string } = Object.fromEntries(
-  Object.entries(KEYWORD_MAP).map(([key, value]) => [value, key]),
-);
-
 export default function ReviewPage() {
-  const params = useParams();
-  const searchParams = useSearchParams();
-  const router = useRouter();
-  const queryClient = useQueryClient();
-
-  const initialContent = searchParams.get("content") || "";
-  const initialKeywordsParam = searchParams.get("keywords") || "";
-  const reviewId = searchParams.get("reviewId");
-
-  const isEditMode = !!reviewId;
-  const title = isEditMode ? "방문 후기 수정" : "방문 후기";
-
-  const getInitialFilters = () => {
-    if (!initialKeywordsParam) return [];
-    return initialKeywordsParam.split(",");
-  };
-
-  const [selectedFilters, setSelectedFilters] = useState<string[]>(getInitialFilters());
-
-  const [review, setReview] = useState(initialContent);
-  const [isModalOpen, setIsModalOpen] = useState(false);
-
-  const textOnlyFilters = REVIEW_FILTERS.map((filter) => filter.split(" ").slice(1).join(" "));
-
-  const createReviewMutation = useMutation({
-    mutationFn: createReview,
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ["reviews", params.id] });
-      setIsModalOpen(true);
-    },
-  });
-
-  const updateReviewMutation = useMutation({
-    mutationFn: updateReview,
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ["reviews", params.id] });
-      setIsModalOpen(true);
-    },
-  });
-
-  const handleFilterChange = (items: string[]) => {
-    const englishKeys = items.map((itemText) => {
-      const fullFilter = REVIEW_FILTERS.find((f) => f.includes(itemText));
-      return fullFilter ? KEYWORD_MAP[fullFilter] : "";
-    });
-    setSelectedFilters(englishKeys.filter(Boolean));
-  };
-
-  const selectedTextOnlyFilters = selectedFilters
-    .map((key) => {
-      const fullFilter = REVERSE_KEYWORD_MAP[key];
-      return fullFilter ? fullFilter.split(" ").slice(1).join(" ") : "";
-    })
-    .filter(Boolean);
-
-  const handleSubmit = () => {
-    if (isEditMode) {
-      updateReviewMutation.mutate({
-        reviewId: Number(reviewId),
-        review: { content: review, keywords: selectedFilters },
-      });
-    } else {
-      createReviewMutation.mutate({
-        storeId: Number(params.id),
-        content: review,
-        keywords: selectedFilters,
-      });
-    }
-  };
+  const {
+    title,
+    isEditMode,
+    selectedFilters,
+    review,
+    setReview,
+    isModalOpen,
+    setIsModalOpen,
+    textOnlyFilters,
+    handleFilterChange,
+    selectedTextOnlyFilters,
+    handleSubmit,
+    handleModalClose,
+  } = useReviewForm();
 
   return (
     <TransitionLayout dynamicTitle={title}>
@@ -114,7 +51,7 @@ export default function ReviewPage() {
             isModalOpen={isModalOpen}
             setIsModalOpen={setIsModalOpen}
             isEditMode={isEditMode}
-            onClose={() => router.push(`/store/${params.id}`)}
+            onClose={handleModalClose}
           />
         )}
       </div>
