+++ src/entities/review/model/constants.ts
@@ -1,4 +1,4 @@
-export const REVIEW_KEYWORD_MAP: Record<string, string> = {
+export const REVIEW_KEYWORD_EN_KO_MAP: Record<string, string> = {
   GOOD_FOR_SOLO: "🏃‍♂️ 빠른 회전률",
   OPEN_ATMOSPHERE: "✨ 넓은 매장",
   KIND_SERVICE: "😊 친절한 응대",
@@ -6,8 +6,8 @@ export const REVIEW_KEYWORD_MAP: Record<string, string> = {
   BEST_TASTE: "👍 보증된 맛",
 };
 
-export const REVIEW_FILTERS = Object.values(REVIEW_KEYWORD_MAP);
+export const REVIEW_FILTERS = Object.values(REVIEW_KEYWORD_EN_KO_MAP);
 
-export const KEYWORD_MAP: { [key: string]: string } = Object.fromEntries(
-  Object.entries(REVIEW_KEYWORD_MAP).map(([key, value]) => [value, key]),
+export const REVIEW_KEYWORD_KO_EN_MAP: { [key: string]: string } = Object.fromEntries(
+  Object.entries(REVIEW_KEYWORD_EN_KO_MAP).map(([key, value]) => [value, key]),
 );
